Galaxy: fix repeating galaxy, stop wasting memory
- DensityMapGalaxy did not clamp texture accesses, resulting in out-of-bounds reads that manifested as repeating galaxies if the map was scrolled far enough
- 8bpp grayscale image data was stored in a 32-bit float only to be read and returned as 8bpp texture data.
- The texture read code will fail if the galaxy_dense image is anything other than a single-channel grayscale bitmap. Added an explicit error to catch that case.

+++ src/galaxy/Galaxy.h
@@ -79,7 +79,7 @@ public:
 	virtual Uint8 GetSectorDensity(const int sx, const int sy, const int sz) const;
 
 private:
-	std::unique_ptr<float[]> m_galaxyMap;
+	std::unique_ptr<uint8_t[]> m_galaxyMap;
 	Sint32 m_mapWidth, m_mapHeight;
 };
